+++ src/main/java/ph/devcon/rapidpass/entities/ScannerDevice.java
@@ -35,7 +35,6 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 /**
- *
  * @author eric
  */
 @Entity
@@ -71,12 +70,12 @@ public class ScannerDevice implements Serializable {
     @Column(name = "date_time_updated")
     @Temporal(TemporalType.TIMESTAMP)
     private Date dateTimeUpdated;
-//    @JoinColumn(name = "registrar_id", referencedColumnName = "id")
+    //    @JoinColumn(name = "registrar_id", referencedColumnName = "id")
 //    @ManyToOne
 //    private Registrar registrarId;
     @Column(name = "registrar_id")
     private Integer registrarId;
-//    @JoinColumn(name = "registrar_user_id", referencedColumnName = "id")
+    //    @JoinColumn(name = "registrar_user_id", referencedColumnName = "id")
 //    @ManyToOne
 //    private RegistrarUser registrarUserId;
     @Column(name = "registrar_user_id")
@@ -109,5 +108,5 @@ public class ScannerDevice implements Serializable {
     public String toString() {
         return "ph.devcon.rapidpass.entities.ScannerDevice[ id=" + id + " ]";
     }
-    
+
 }